Initial 1.20 port

+++ src/main/java/galena/oreganized/client/particle/CustomDrippingParticle.java
@@ -50,7 +50,7 @@ public class CustomDrippingParticle extends TextureSheetParticle {
                 this.xd *= (double)0.98F;
                 this.yd *= (double)0.98F;
                 this.zd *= (double)0.98F;
-                BlockPos blockpos = new BlockPos(this.x, this.y, this.z);
+                BlockPos blockpos = new BlockPos((int) this.x, (int) this.y, (int) this.z);
                 FluidState fluidstate = this.level.getFluidState(blockpos);
                 if (fluidstate.getType() == this.type && this.y < (double)((float)blockpos.getY() + fluidstate.getHeight(this.level, blockpos))) {
                     this.remove();
